+++ Core/ConnectionSML/src/sock_Socket.cpp
@@ -209,7 +209,7 @@ bool Socket::SendBuffer(char const* pSendBuffer, uint32_t bufferSize)
         return false;
     }
     
-    uint32_t bytesSent = 0 ;
+    size_t bytesSent = 0 ;
     size_t   thisSend = 0 ;
     
     // May need repeated calls to send all of the data.
@@ -361,7 +361,7 @@ bool Socket::ReceiveBuffer(char* pRecvBuffer, uint32_t bufferSize)
         return false;
     }
     
-    uint32_t bytesRead = 0 ;
+    size_t bytesRead = 0 ;
     size_t   thisRead  = 0 ;
     
     // Check our incoming data is valid
